Added types for API response, organized page files, components, and utils. Cleaned up repititive code
