+++ ui/src/Pages/Login.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 
 const AUTH_URL = `https://accounts.spotify.com/authorize?client_id=${process.env.REACT_APP_CLIENT_ID}&response_type=code&redirect_uri=${process.env.REACT_APP_REDIRECT_URI}&scope=streaming%20user-read-email%20user-read-private%20user-library-read%20user-library-modify%20user-read-playback-state%20user-modify-playback-state%20playlist-read-private%20playlist-read-collaborative%20user-top-read`;
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export default function Login() {
   return (
     <div className="flex items-center justify-center h-screen py-12 px-4 sm:px-6 lg:px-8 bg-black">
