+++ ui/src/components/NavBar.tsx
@@ -1,8 +1,7 @@
 import React from "react";
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export default function NavBar() {
   return (
-    <nav className="bg-zinc-900">
+    <nav className="bg-black">
       <div className="mx-auto max-w-8xl px-2 sm:px-6 lg:px-8">
         <div className="relative flex h-16 items-center justify-between">
           <div className="flex-items">
@@ -23,21 +22,21 @@ export default function NavBar() {
                 <div className="flex space-x-4">
                   <a
                     href="http://localhost:3000/dashboard"
-                    className="bg-zinc-800 text-white rounded-md px-3 py-2 text-sm font-medium"
+                    className="bg-zinc-900 text-white rounded-md px-3 py-2 text-sm font-medium"
                     aria-current="page"
                   >
                     Dashboard
                   </a>
                   <a
                     href="http://localhost:3000/search"
-                    className="bg-zinc-800 text-white rounded-md px-3 py-2 text-sm font-medium"
+                    className="bg-zinc-900 text-white rounded-md px-3 py-2 text-sm font-medium"
                     aria-current="page"
                   >
                     Search
                   </a>
                   <a
                     href="http://localhost:3000/library"
-                    className="bg-zinc-800 text-white rounded-md px-3 py-2 text-sm font-medium"
+                    className="bg-zinc-900 text-white rounded-md px-3 py-2 text-sm font-medium"
                     aria-current="page"
                   >
                     My Library
