+++ ui/src/types.ts
@@ -1,7 +1,5 @@
 export type Image = {
   url: string;
-  height: number;
-  width: number;
 };
 
 export type Artist = {
@@ -9,6 +7,8 @@ export type Artist = {
   images: Image[];
   name: string;
   popularity: number;
+  items: Artist[];
+  tracks: Track[];
 };
 
 export type Track = {
@@ -20,37 +20,59 @@ export type Track = {
   duration_ms: number;
   track_number: number;
   explicit: boolean;
-
   release_date: string;
-
   description: string;
+  items: Track[];
+  tracks: Track[];
+  track: Track;
 };
 
 export type Album = {
+  copyrights: string[];
+  total_tracks: string;
   id: string;
   images: Image[];
   name: string;
-  length: number;
   release_date: string;
   artists: Artist[];
+  items: Album[];
+  tracks: Track;
 };
 
 export type Playlist = {
+  tracks: Track;
   id: string;
   images: Image[];
   name: string;
-  length: number;
   release_date: string;
-
   description: string;
+  items: Playlist[];
+  owner: Owner;
+  followers: Followers;
 };
 
-export type PlaylistTracks = {
+export type PlaylistTrack = {
   id: string;
   images: Image[];
   name: string;
   release_date: string;
   track: Track;
-
+  artists: Artist[];
   description: string;
+};
+
+export type API = {
+  albums: Album;
+  playlists: Playlist;
+  artists: Artist;
+  tracks: Track;
+  message: string;
+};
+
+export type Owner = {
+  display_name: string;
+};
+
+export type Followers = {
+  total: string;
 };
