+++ ui/tailwind.config.js
@@ -1,14 +1,7 @@
 module.exports = {
+    important: true,
     content: [
         "./src/**/*.{js,jsx,ts,tsx}",
-    ],
-
-    important: '#root',
-    theme: {
-        extend: {
-            height: {'100': '26rem'}, width: {'100': '26rem'}
-        },
-    },
-    plugins: [],
+    ]
 }
